+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Add Polychromatic GUI for managing Razer accessories.
 - Change the container build to support all image types.
 - Add a GitHub Actions CI pipeline for Manjaro image builds.
+- Change the GRUB menu to show winesapOS as the distribution name.
 
 ## 4.0.0 - The Conversion Update